fix: cast idle_hours to int in cleanup_stale_sessions SQL
Postgres make_interval(hours => ...) expects integer, not bigint.
The i64 bind caused a type mismatch warning on every startup.

+++ src/queries/session_queries.rs
@@ -307,7 +307,7 @@ impl SessionRepository {
             UPDATE scan_sessions
             SET ended_at = NOW()
             WHERE ended_at IS NULL
-              AND started_at < NOW() - make_interval(hours => $1)
+              AND started_at < NOW() - make_interval(hours => $1::int)
             RETURNING id
             "#,
         )
